replace mimir_cluster lib with a custom module
